% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-test.R
\docType{package}
\name{test-package}
\alias{test}
\alias{test-package}
\title{test: Just a test}
\description{
Nothing to see here.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpadge/pkgdown-test}
  \item Report bugs at \url{https://github.com/mpadge/pkgdown-test/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
